scroll view with key board
In the Adopt view, the user must fill some textfields.
In previous version, there were two diferents view, one for name, the
other for email and phone number. When we wanted to write into a field,
we could not always see what we were writing, because of keyboard
appearing.
Now there is only one view withe three fields and the view move up when
the keyboard appears.